Refactor
Moved axios service file from services to shared/third-party
